Adapt the search/TestPhraseQuery to test PimPhraseQuery
Created a copy of the test where a PimIndexWriter is created instead
of an IndexWriter, and PimPhraseQuery are used instead of PhraseQuery.
A PimPhraseQuery is executed using the PIM index only if the slop is zero,
as there is currently no support for slop in the PIM Index search
algorithm. Otherwise it is rewritten as a PhraseQuery.