Redesign of the PimSystemManager to handle batches of queries
Queries are written in parallel in a queue of byte buffers.
A thread regularly checks the queue and send a batch of queries to the
PIM system, retrieves the results and make them available to the client
searcher threads.
No PIM HW is used for now, the query is still handled by the
software model running on the PIM index.